the basic works. now lump

+++ src/Constraint.h
@@ -62,7 +62,7 @@ public:
 	int idxER;								// Reduced equality constraint indices
 	int idxIM;								// Maximal inequality constraint indices
 	int idxIR;								// Reduced inequality constraint indices
-	Eigen::MatrixXd idxQ;					// Associated DOF indices
+	Eigen::MatrixXi idxQ;					// Associated DOF indices
 	int nidxQ;
 
 	bool activeM;							// Whether the maximal inequality constraint is active
