+++ src/Joint.cpp
@@ -393,7 +393,7 @@ void Joint::draw_(shared_ptr<MatrixStack> MV, const shared_ptr<Program> prog, co
 		if (presc->activeER) {
 			alpha = 2.0;
 		}
-		MV->scale(alpha * m_draw_radius);
+		MV->scale(static_cast<float>(alpha * m_draw_radius));
 		glUniformMatrix4fv(prog->getUniform("MV"), 1, GL_FALSE, glm::value_ptr(MV->topMatrix()));
 		m_jointShape->draw(prog);
 		MV->popMatrix();
